Move create and save routine to own screens.
